Moved convertToInteger into own utils file. Added updated TagManager.dataLayer code to JoinChallengeButton based on discussions in Analytics team meeting.
