+++ package.json
@@ -142,7 +142,7 @@
     "react": "~17.0.2",
     "react-bootstrap": "^1.6.4",
     "react-confetti": "^6.1.0",
-    "react-copy-to-clipboard": "^5.0.3",
+    "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "~17.0.2",
     "react-fullstory": "^1.4.0",
     "react-ga4": "^2.1.0",
