+++ src/js/common/utils/convertToInteger.js
@@ -1,0 +1,3 @@
+export default function convertToInteger (incomingNumber) {
+  return parseInt(incomingNumber, 10) || 0;
+}
