+++ src/js/components/Widgets/ItemActionBar/ItemActionBar.jsx
@@ -15,13 +15,13 @@ import VoterActions from '../../../actions/VoterActions';
 import DesignTokenColors from '../../../common/components/Style/DesignTokenColors';
 import { openSnackbar } from '../../../common/components/Widgets/SnackNotifier';
 import AppObservableStore from '../../../common/stores/AppObservableStore';
+import convertToInteger from '../../../common/utils/convertToInteger';
 import { isWebApp } from '../../../common/utils/isCordovaOrWebApp';
 import isMobileScreenSize from '../../../common/utils/isMobileScreenSize';
 import Cookies from '../../../common/utils/js-cookie/Cookies';
 import { renderLog } from '../../../common/utils/logging';
 import normalizedImagePath from '../../../common/utils/normalizedImagePath';
 import stringContains from '../../../common/utils/stringContains';
-import { convertToInteger } from '../../../common/utils/textFormat';
 import webAppConfig from '../../../config';
 import VoterConstants from '../../../constants/VoterConstants';
 import SupportStore from '../../../stores/SupportStore';
